# server/swagger.yaml
openapi: 3.0.0
info:
  title: FloNeo LCNC Platform API
  description: RESTful API for the FloNeo Low-Code No-Code Platform
  version: 1.0.0
  contact:
    name: FloNeo Team
    email: support@floneo.com
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/floneo-ai/status:
    get:
      summary: Get Floneo AI system status
      description: Returns the current status and capabilities of the Floneo AI system
      tags:
        - Floneo AI
      security:
        - BearerAuth: []
      responses:
        '200':
          description: AI system status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        example: "active"
                      version:
                        type: string
                        example: "1.0.0"
                      totalIdeas:
                        type: integer
                        example: 5
        '401':
          description: Unauthorized
  /api/floneo-ai/ideas:
    get:
      summary: Get AI-generated app ideas
      description: Returns AI-suggested app ideas with recommended templates
      tags:
        - Floneo AI
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [Insurance, CRM, Forms, E-commerce, Dashboards]
          description: Filter ideas by category
        - in: query
          name: complexity
          schema:
            type: string
            enum: [Low, Medium, High]
          description: Filter ideas by complexity level
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
          description: Number of ideas to return
      responses:
        '200':
          description: AI ideas retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      ideas:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                            description:
                              type: string
                            suggestedTemplate:
                              type: string
                            templateId:
                              type: integer
                            category:
                              type: string
                            complexity:
                              type: string
                      totalIdeas:
                        type: integer
        '401':
          description: Unauthorized

  # ===== CANVAS APIS =====
...